requirements.txt
# Consolidated Python dependencies for all services
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-dateutil==2.9.0.post0
httpx==0.28.1
feedparser==6.0.11
APScheduler==3.11.0
langgraph==1.0.10
langchain-core==1.2.23
langchain-nvidia-ai-endpoints==1.2.1
redis[hiredis]>=5,<6
openai>=1,<2
sentence-transformers>=3,<5
faiss-cpu>=1.8,<2
psycopg2-binary>=2.9,<3
python-dotenv>=1,<2
python-multipart>=0.0.9,<1
cryptography>=42,<46
